customcluster: refactor to reduce file size of controller.go
