+++ pkg/controllers/customcluster_clusterconfig.template
@@ -1,0 +1,7 @@
+kube_version: {{ .KubeVersion}}
+download_run_once: true
+download_container: false
+download_localhost: true
+# network
+kube_pods_subnet: {{ .PodCIDR }}
+kube_network_plugin: {{ .CNIType }}
